project/backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10

# Database
pymongo==4.6.1
mongoengine==0.28.2

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# Security
werkzeug==3.0.1

# Validation
marshmallow==3.20.1
bleach==6.1.0

# Utilities
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0
